refactor(util): 숫자 관련 파싱 기능을 NumberParser 에서 통합 관리하도록 변경
- PurchaseAmountParser 클래스 삭제
- 금액 관련 검증 기능을 PurchaseAmount VO에 위임
